Formally address FreeTDS 0.64 compiler warnings.
* Harmonize type usage, mainly addressing 64-bit cleanliness warnings
  and problematic mixing of signed and unsigned types; add explicit
  casts where safe to do so.
* Comment out unused variables and assignments thereto.
* Arrange for MD4 and MD5 calculations to use a union to legimitize
  their aliasing.
* Keep static assertions warning-free by using _Static_assert where
  available, and the standard offsetof macro from <stddef.h>.
* ctlib.h: #define _CS_CURS_TYPE_* to TDS_CURSOR_STATE_* rather than
  (typelessly) duplicating their values.
* dblib/dblib.c: Express -2?\194?\179?\194?\185 in a manner all compilers accept happily.
* tds/gssapi.c: Suppress GSSAPI deprecation warnings on Mac OS X.
* tds/sspi.c: Comment out code that used a deprecated function to check
  for ancient Windows versions, on which its modern substitute isn't
  available in the first place.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@70299 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/dbapi/driver/ftds64/freetds/hmac_md5.h
@@ -13,19 +13,19 @@ typedef struct
 } HMACMD5Context;
 
 void hmac_md5_init_limK_to_64(const unsigned char* key,
-                              int key_len,
+                              size_t key_len,
                               HMACMD5Context *ctx);
 void hmac_md5_update(const unsigned char* text,
-                     int text_len,
+                     size_t text_len,
                      HMACMD5Context *ctx);
 void hmac_md5_final(unsigned char* digest,
                     HMACMD5Context *ctx);
 void hmac_md5(const unsigned char key[16],
               const unsigned char* data,
-              int data_len,
+              size_t data_len,
               unsigned char* digest);
 void hmac_md5_init_rfc2104(const unsigned char* key,
-                           int key_len,
+                           size_t key_len,
                            HMACMD5Context *ctx);
 
 #endif /* !HMAC_MD5_H */
