+++ src/dbapi/driver/ftds64/freetds/tds/hmac_md5.c
@@ -8,7 +8,7 @@
  the rfc 2104 version of hmac_md5 initialisation.
 ***********************************************************************/
 void hmac_md5_init_rfc2104(const unsigned char* key,
-                           int key_len,
+                           size_t key_len,
                            HMACMD5Context *ctx)
 {
     int i;
@@ -47,7 +47,7 @@ void hmac_md5_init_rfc2104(const unsigned char* key,
  the microsoft version of hmac_md5 initialisation.
 ***********************************************************************/
 void hmac_md5_init_limK_to_64(const unsigned char* key,
-                              int key_len,
+                              size_t key_len,
                               HMACMD5Context *ctx)
 {
     /* if key is longer than 64 bytes truncate it */
@@ -62,7 +62,7 @@ void hmac_md5_init_limK_to_64(const unsigned char* key,
  update hmac_md5 "inner" buffer
 ***********************************************************************/
 void hmac_md5_update(const unsigned char* text,
-                     int text_len,
+                     size_t text_len,
                      HMACMD5Context *ctx)
 {
     MD5Update(&ctx->ctx, text, text_len); /* then text of datagram */
@@ -89,7 +89,7 @@ void hmac_md5_final(unsigned char* digest, HMACMD5Context *ctx)
 ************************************************************/
 void hmac_md5(const unsigned char key[16],
               const unsigned char* data,
-              int data_len,
+              size_t data_len,
               unsigned char* digest)
 {
     HMACMD5Context ctx;
